Make end menu unique
Fix the "ñapa" for the jam. The end menu had a child
start menu and we passed parameters to configure its
text and behavior when it was teh end menu.

Now both scenes hardcode their text, the music used
and the behavior.

Also turned the level array into a resource stored in
a file, so it isn't duplicated.